backend/src/auth/requirements.txt
alembic==1.9.1
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==3.4
importlib-metadata==5.2.0
importlib-resources==5.10.1
isort==5.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
mypy-extensions==0.4.3
mysqlclient==2.1.1
pathspec==0.10.3
pika==1.3.1
platformdirs==2.6.1
PyJWT==2.6.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.45
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
